refactor motd formatter classes
