+++ apps/dashboard/app/models/batch_connect/app.rb
@@ -1,4 +1,4 @@
- require "smart_attributes"
+require "smart_attributes"
 
 module BatchConnect
   class App
