--- apps/dashboard/app/models/motd_formatter_plaintext.rb
@@ -1,14 +1,0 @@
-class MotdFormatterPlaintext
-
-  attr_reader :content, :title
-
-  def initialize(motd_file)
-    motd_file = MotdFile.new unless motd_file
-    @title = motd_file.title
-    @content = motd_file.content
-  end
-
-  def to_partial_path
-    "dashboard/motd_plaintext"
-  end
-end
