+++ apps/dashboard/test/models/motd_formatter/motd_formatter_osc_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class MotdTest < ActiveSupport::TestCase
+class MotdFormatter::OscTest < ActiveSupport::TestCase
   # test date order of example Motd file
   #
   test "motd message date format" do
@@ -9,22 +9,22 @@ class MotdTest < ActiveSupport::TestCase
 
     # assume year month day
     msg = "2016/05/04\n--- NEW CLUSTER\n\nSomething good!"
-    assert_equal date, MotdFormatterOsc::Message.from(msg).date
+    assert_equal date, MotdFormatter::Osc::Message.from(msg).date
     msg = "2016-05-04\n--- NEW CLUSTER\n\nSomething good!"
-    assert_equal date, MotdFormatterOsc::Message.from(msg).date
+    assert_equal date, MotdFormatter::Osc::Message.from(msg).date
     msg = "2016.05.04\n--- NEW CLUSTER\n\nSomething good!"
-    assert_equal date, MotdFormatterOsc::Message.from(msg).date
+    assert_equal date, MotdFormatter::Osc::Message.from(msg).date
     msg = "2016 05 04\n--- NEW CLUSTER\n\nSomething good!"
-    assert_nil MotdFormatterOsc::Message.from(msg)
+    assert_nil MotdFormatter::Osc::Message.from(msg)
     msg = "2016+05+04\n--- NEW CLUSTER\n\nSomething good!"
-    assert_nil MotdFormatterOsc::Message.from(msg)
+    assert_nil MotdFormatter::Osc::Message.from(msg)
   end
 
   test "test when motd formatter_osc_valid" do
 
     path = "#{Rails.root}/test/fixtures/files/motd_valid"
     motd_file = MotdFile.new(path)
-    formatted_motd = MotdFormatterOsc.new motd_file
+    formatted_motd = MotdFormatter::Osc.new motd_file
     expected_file = File.open(path).read
 
     assert_equal true, motd_file.exist?
@@ -36,7 +36,7 @@ class MotdTest < ActiveSupport::TestCase
   test "test when motd_formatter_osc empty" do
     path = "#{Rails.root}/test/fixtures/files/motd_empty"
     motd_file = MotdFile.new(path)
-    formatted_motd = MotdFormatterOsc.new motd_file
+    formatted_motd = MotdFormatter::Osc.new motd_file
 
     assert_equal true, motd_file.exist?
     assert_equal path, motd_file.motd_path
@@ -47,7 +47,7 @@ class MotdTest < ActiveSupport::TestCase
   test "test when motd formatter_osc_missing" do
     path = "#{Rails.root}/test/fixtures/files/motd_missing"
     motd_file = MotdFile.new(path)
-    formatted_motd = MotdFormatterOsc.new motd_file
+    formatted_motd = MotdFormatter::Osc.new motd_file
 
     assert_equal false, motd_file.exist?
     assert_equal path, motd_file.motd_path
@@ -57,7 +57,7 @@ class MotdTest < ActiveSupport::TestCase
 
   test "test when motd formatter_osc_nil" do
     motd_file = nil
-    formatted_motd = MotdFormatterOsc.new motd_file
+    formatted_motd = MotdFormatter::Osc.new motd_file
 
     assert_not_nil formatted_motd.content
   end
